+++ src/main/Yardarm/Generation/Tag/TagImplementationCategory.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.OpenApi.Models;
+
+namespace Yardarm.Generation.Tag
+{
+    /// <summary>
+    /// Marks type generators for the concrete class implementation of a <see cref="OpenApiTag"/>.
+    /// </summary>
+    public class TagImplementationCategory
+    {
+        private TagImplementationCategory()
+        {
+        }
+    }
+}
